+++ programs/cream-pad/src/states/collection/collection_auction_account.rs
@@ -37,7 +37,7 @@ pub struct CollectionAuctionAccount {
 
     pub current_round: u16,
 
-    pub boost_history: Vec<u64>,
+    pub boost_history: Vec<f64>,
 
     pub decay_model: DecayModelType,
