+++ programs/cream-pad/src/states/collection/collection_auction_round_account.rs
@@ -18,7 +18,7 @@ pub struct CollectionAuctionRoundAccount {
 
     pub total_user_count: u64,
 
-    pub boost: u64,
+    pub boost: f64,
 
     pub price: u64,
